Silence unused-param warnings and minor fixes
Add explicit (void) casts for unused parameters in several macOS files to silence compiler warnings (macos_menu.m, macos_system.m, macos_window.m). Fix a comment in unix_joystick.c. Correct an unsigned sentinel comparison in unix_shared.c by changing if (keynum == -1) to if (keynum == (uint32_t)-1) to avoid incorrect signed/unsigned comparison.

+++ unix/unix_shared.c
@@ -149,7 +149,7 @@ axKeynumToString(uint32_t keynum)
   keyname_t* kn;
   static char tinystr[2];
   keynum = keynum&0xff;
-  if (keynum == -1)
+  if (keynum == (uint32_t)-1)
     return "<KEY NOT FOUND>";
   if (keynum > 32 && keynum < 127) { // printable ascii
     tinystr[0] = keynum;
